nono

+++ length_utils.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   length_utils.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dongguki <dongguki@student.42seoul.kr>     +#+  +:+       +#+        */
+/*   By: dongguki <dongguki@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/01 01:44:21 by dongguki          #+#    #+#             */
-/*   Updated: 2021/02/01 01:44:21 by dongguki         ###   ########.fr       */
+/*   Updated: 2021/02/02 15:16:46 by dongguki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int		lengthplus(t_cfc cfc, unsigned long val)
+int		lengthplus(t_cfc cfc, unsigned long long val)
 {
 	int	i;
